requirements.txt
Flask==2.1.0
Flask-Cors==3.0.10
mysql-connector-python
scrypt
flask_sqlalchemy
Flask-APScheduler
tzlocal==2.1
pyjwt